Implement Kafka integration for product image handling; add image upload and deletion services; update application configuration

+++ backend/product-service/src/main/java/com/ecommerce/product_service/dto/ProductImageDto.java
@@ -19,7 +19,7 @@ public class ProductImageDto implements Serializable {
 
     int id;
     int productId;
-    int orderIndex;
+    byte orderIndex;
     MultipartFile file;
 
     public static ProductImageDto fromEntity(com.ecommerce.product_service.entity.ProductImage productImage) {
